Iterated through the data in The Datajs file, created small boxes in a big box.The small boxes contain the info of one element in the data array. Placed an icon next to the title,but i traeted it as button.
On click the icon changes,so i decided to create an array of two icons.
eg const icons = [<icon1/>,<icon2/>],the are both components i got from a
library called React Icons .so when i render the data.map() function
i gave the <button>{icon[index]}<button> solution since it was easy and
quick to reference the icon which i chocse when i click.

Then i toggled and displayed the data .